requirements.txt
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
streamlit>=1.28.0